+++ CHANGES.rst
@@ -14,6 +14,7 @@ Major Changes
 
 - feat(cguard): refactor distinct behaviors into new policy types
 - refactor(cguard): split interfaces into separate modules
+- feat(cguard): automate ref cleanup, improve logic, docs & tests
 
 
 Minor Changes